[27646] Calcul de la taille des docservers si date de dernière MAJ vide

See merge request maarch/MaarchCourrier!1399

+++ src/app/docserver/controllers/DocserverController.php
@@ -194,7 +194,7 @@ class DocserverController
             $now = new DateTime();
             $hoursDiff = ($now->diff($lastCalculationDateTime)->days * 24) + $now->diff($lastCalculationDateTime)->h;
 
-            if ($hoursDiff < 2) {
+            if (!empty($last_calculation_date['param_value_date']) && $hoursDiff < 2) {
                 unlink($lockFile);
                 return $response->withStatus(403)->withJson(['error' => 'Last calculation is too early']);
             }
